Test Updates - Added special cases to help menu test (to reflect differences in the program) - Removed extra flags from lexer - updated test_help to test_lexer
