+++ src/umleditor/mvc_controller/uml_lexer.py
@@ -7,7 +7,7 @@ _command_flag_map = {
     "list"  : ["a","c","r","d"],
     "fld"   : ["a","d","r"],
     "mthd"  : ["a","d","r"],
-    "rel"   : ["a","t","d","e"],
+    "rel"   : ["a","t","d"],
     "save"  : [""],
     "load"  : [""],
     "exit"  : [""],
@@ -22,7 +22,7 @@ _command_function_map = {
     "list"  : ["list_everything","list_entities","list_relations","list_entity_details"],
     "fld"   : ["add_field","delete_field","rename_field"],
     "mthd"  : ["add_method","delete_method","rename_method"],
-    "rel"   : ["add_relation","change_relation_type","delete_relation","edit_relation"],
+    "rel"   : ["add_relation","change_relation_type","delete_relation"],
     "save"  : ["save"],
     "load"  : ["load"],
     "exit"  : ["quit"],
